+++ src/domain/interfaces/IAuthJwtCookieConfig.ts
@@ -1,0 +1,7 @@
+export interface IAuthJwtCookieConfig {
+    signed?: boolean,
+    path?: string,
+    domain?: string,
+    secure?: boolean,
+    sameSite?: boolean | 'lax' | 'strict' | 'none',
+}