+++ src/infrastructure/config.ts
@@ -1,4 +1,5 @@
 import {AuthConfirmProviderTypeEnum} from '../domain/enums/AuthConfirmProviderTypeEnum';
+import {IAuthJwtCookieConfig} from '../domain/interfaces/IAuthJwtCookieConfig';
 
 export interface IAuthModuleConfig {
     jwtAccessSecretKey?: string,
@@ -7,6 +8,7 @@ export interface IAuthModuleConfig {
     refreshTokenExpiresSec?: string, // Additional token expiration time for FilesAuthGuard
     filesTokenAdditionalTime?: string,
     confirm: IAuthConfirmConfig,
+    jwtCookie?: IAuthJwtCookieConfig,
 }
 
 export interface IAuthConfirmConfig {
@@ -35,5 +37,10 @@ export default () => ({
         providerName: process.env.AUTH_PROVIDER_NAME || 'smsc',
         providerType: process.env.AUTH_PROVIDER_TYPE || AuthConfirmProviderTypeEnum.VOICE, // or "sms", or "call"
         messageTemplate: process.env.AUTH_CONFIRM_MESSAGE_TEMPLATE || 'Ваш код авторизации в {appTitle} - {code}',
+    },
+    jwtCookie: {
+        secure: process.env.APP_ENVIRONMENT !== 'dev',
+        sameSite: 'lax',
+        path: '/',
     },
 } as IAuthModuleConfig);